# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Springbok Finance Portal API
  version: 1.0.0
  description: Production-ready API for purchase order and invoice management
  contact:
    name: Springbok Development Team
    email: dev@springbok.com

servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://api.springbok.com/api
    description: Production

security:
  - bearerAuth: []

paths:
  /auth/login:
    post:
      summary: User login
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
              required: [email, password]
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
        '400':
          description: Validation error

  /auth/logout:
    post:
      summary: User logout
      tags: [Authentication]
      responses:
        '200':
          description: Logged out successfully
        '401':
          description: Not authenticated

  /purchase-orders:
    get:
      summary: Get purchase orders
      tags: [Purchase Orders]
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, ACTIVE, COMPLETED, CANCELLED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  purchase_orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/PurchaseOrder'
                  count:
                    type: integer
        '401':
          description: Not authenticated

    post:
      summary: Create purchase order
      tags: [Purchase Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                po_number:
                  type: string
                  pattern: '^PO-\d{4}-\d{3}$'
                vendor_id:
                  type: string
                  format: uuid
                department:
                  type: string
                description:
                  type: string
                  minLength: 10
                amount:
                  type: number
                  format: decimal
                expected_completion_date:
                  type: string
                  format: date
                reporting_period:
                  type: string
                  pattern: '^\d{4}-\d{2}$'
              required: [po_number, vendor_id, department, description, amount, expected_completion_date, reporting_period]
      responses:
        '201':
          description: Purchase order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  purchase_order:
                    $ref: '#/components/schemas/PurchaseOrder'
        '403':
          description: Insufficient permissions
        '409':
          description: PO number already exists

  /invoices:
    get:
      summary: Get invoices
      tags: [Invoices]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [RECEIVED, APPROVED, PAID, REJECTED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'

    post:
      summary: Record invoice
      tags: [Invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invoice_number:
                  type: string
                  pattern: '^[A-Z0-9]{5,20}$'
                po_id:
                  type: string
                  format: uuid
                vendor_id:
                  type: string
                  format: uuid
                amount:
                  type: number
                  format: decimal
              required: [invoice_number, po_id, vendor_id, amount]
      responses:
        '201':
          description: Invoice recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoice:
                    $ref: '#/components/schemas/Invoice'
        '403':
          description: Finance role required

  /invoices/{id}/approve:
    post:
      summary: Approve invoice
      tags: [Invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                approval_notes:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Invoice approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoice:
                    $ref: '#/components/schemas/Invoice'

  /invoices/{id}/mark-paid:
    post:
      summary: Mark invoice as paid
      tags: [Invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_method:
                  type: string
                  enum: [BANK_TRANSFER, CHEQUE, CARD]
                reference:
                  type: string
                  maxLength: 100
              required: [payment_method, reference]
      responses:
        '200':
          description: Invoice marked as paid

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [po_owner, finance, senior_finance, receiver]
        department:
          type: string

    PurchaseOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        po_number:
          type: string
        vendor_id:
          type: string
          format: uuid
        department:
          type: string
        description:
          type: string
        amount:
          type: number
          format: decimal
        status:
          type: string
          enum: [DRAFT, ACTIVE, COMPLETED, CANCELLED]
        created_at:
          type: string
          format: date-time

    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoice_number:
          type: string
        po_id:
          type: string
          format: uuid
        vendor_id:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
        status:
          type: string
          enum: [RECEIVED, APPROVED, PAID, REJECTED]
        created_at:
          type: string
          format: date-time